+++ src/components/EventScheduleItem/index.js
@@ -1,0 +1,1 @@
+export { default as EventScheduleItem } from './EventScheduleItem';
